More file/folder restructuring in client
