+++ client/src/components/ClassInput.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
-import { Link } from 'react-router-dom';
-import "./ClassInput.css";
+import { Link } from "react-router-dom";
+import "./styles/ClassInput.css";
 import "rc-time-picker/assets/index.css";
 import TimePicker from "rc-time-picker";
 
@@ -462,7 +462,9 @@ class ClassInput extends Component {
               </div>
             </form>
           </div>
-		  <li><Link to='/signin'>Account Page</Link></li>
+          <li>
+            <Link to="/signin">Account Page</Link>
+          </li>
         </header>
       </div>
     );
